+++ src/main/java/edu/oregonstate/eecs/shp3d/processor/HeightField.java
@@ -1,5 +1,5 @@
 package edu.oregonstate.eecs.shp3d.processor;
 
 interface HeightField {
-	float heightAt(float x, float y);
+	float heightAt(double x, double y);
 }
